pyproject.toml
[tool.poetry]
name = "caldata-mdsa-caltrans-pems"
version = "0.1.0"
description = "\"Traffic Operations PeMS Modernization\""
authors = ["Traffic Operations <Zhenyu.Zhu@dot.ca.gov>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mkdocs-material = "~9.1.3"
dbt-core = "~1.6.0"
dbt-snowflake = "~1.6.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"
sqlfluff = "2.3.0"
sqlfluff-templater-dbt = "2.3.0"


[tool.poetry.group.airflow.dependencies]
apache-airflow = "2.7.2"
fsspec = "2023.9.2"
boto3 = "1.28.17"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
